refactor(#724): move canonical-artist resolution + shared similarity floor out of the orchestrator
Pure code motion, no behavior change (PR 2 of the #722 orchestrator decomposition). The canonical-artist resolution cluster moves verbatim from lookup/orchestrator.py to the new lookup/artist_resolution.py: the LML#504/#506 env-var rollback gates (_ARTIST_IDENTITY_SPLIT_GATE_ENV_VAR, _FALSE_FLAG_VALUES, _artist_identity_split_gate_enabled, _MB_RESCUE_REQUIRE_SONG_MATCH_ENV_VAR, _mb_rescue_song_match_required), _artist_pair_verified, the resolver pre-pass (_resolver_cache, ResolverOutcome, resolve_canonical_artist), and the observability projections (_log_release_resolution_bind, _project_mb_rescue_attrs, _log_resolver_pre_pass, _log_artist_identity_split_gate). _log_album_title_fallback stays behind — its consumers are inside search_compilations_for_track and it moves with that strategy in PR 4b. CANONICAL_ARTIST_SIMILARITY_FLOOR moves to the new core/thresholds.py leaf (zero non-stdlib imports; lookup/ already imports release.musicbrainz_resolver, so the shared constant can live in neither package), and release/musicbrainz_resolver.py drops its keep-in-sync _SIMILARITY_FLOOR duplicate in favor of the leaf import, keeping the false-positive-tracklist rationale at the floor's use site. The orchestrator imports only the names its remaining code consumes (ResolverOutcome included — search_compilations_for_track constructs a no-swap outcome on the flag-off path); no re-exports.
